Clean up project assets

+++ scripts/set-version.sh
@@ -1,5 +1,13 @@
 #!/bin/bash
 
+# Exit immediately if a command exits with a non-zero status.
+set -e
+
+if [ ! -f .asherah-version ]; then
+  echo "Error: The script is meant to be run from the root directory of the asherah-node project" >&2
+  exit 1
+fi
+
 apt-get update
 apt-get install -y jq
